Simplify ResolveNTFunctionPtr (temporary)
I want to test the theory that the issues we are observing here
are actually a race condition.
- The race condtion would be related with 2 operations that
are not thread safe:
1- check/creation of the map
2- search/insert on the map

I would like to air this CL on dev channel for a week and
observe the crash rate.


BUG=11789

Review URL: http://codereview.chromium.org/199052

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2696c9d9aac84965fdbbeaafd5986ea8c49c624e