linux: turn on -Wextra
This is a followup to an earlier change (r38266) which did most of the
warning-related cleanup.  This one just flips the flag, and fixes some
new warnings that crept in during the window while the flag was off.

BUG=34160

Review URL: http://codereview.chromium.org/597023

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 95250e4c742d27fad6e4bb737a8110a4ee63323c