Seccomp sandbox: Add a policy flag to allow file namespace access to be disabled
This allows file namespace access to be turned on for the purpose of
testing, and we use this in some of the tests, but it is disabled by
default.

This synchronises the Chromium copy with r88 in the non-Chromium copy of
seccomp-sandbox.

BUG=none
TEST=make test

Review URL: http://codereview.chromium.org/3248002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 94609b06fbc45c75bfca8fcafc8866ff0fa135bb