Pull seccomp-sandbox in via DEPS rather than using an in-tree copy
This means changes to the sandbox won't have to be committed twice, to 
both trees. 

This is a retry of r57921, which was committed with git-svn and failed 
to remove the "seccomp" directory. This caused problems when trying 
to "svn checkout" to the same location, and the change was reverted. 
This time I will use SVN to commit the change. 

BUG=none 
TEST=smoke test of running chromium with --enable-seccomp-sandbox 
Review URL: http://codereview.chromium.org/3225010

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 808ea578e1597271a1a73de640e9957babb4f9ea