Revert 79618 - Landing for Julien Tinnes, jln@google.com:
---
chroot to /proc instead of /tmp. This gets rid of a lot of unnecessary
complexity and fixes a race condition.

(Original idea from Markus)

The chroot helper will chroot to /proc/self/fdinfo (or /proc/self/fd). This is
pretty safe because access to this directory is protected by the ptrace() check
in the kernel and the helper is privileged.

Moreover, as soon as the helper _exit() and becomes a zombie, the directory
will be empty. Zygote should wait() for us to make everything deterministric.

We also export SBX_HELPER_PID so that Zygote can specifically wait for the
helper. 
---

BUG=76542
R=markus,agl
Review URL: http://codereview.chromium.org/6683056

TBR=cevans@chromium.org
Review URL: http://codereview.chromium.org/6675053

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0e00451fcfc8cb9974abd21f45bec77082ecce44