Don't fork Zygote as a background process
On Linux, with the setuid sandbox, the Zygote would become a background
process of sort because the setuid sandbox would exit.

The problem is that the Chrome process tree would be broken because the
Zygote would be reparented to init.

In turn, this could create issues with the browser not being able to ptrace()
the Zygote if certain kernel restrictions are in place (e.g. Yama).

BUG=125225
TEST=
NOTRY=true


Review URL: https://chromiumcodereview.appspot.com/10447135

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f2aca70141a15aef16e2244c7dc93e58b914e6d0