Add a parameter to the sandbox policy to allow sandboxed process to run outside of a job and wire it to a cmd line flag.
This is needed for running chrome in Citrix or RemoteApp (Terminal Services) environments.
These envoronments both start the main process inside a job spawned by rdpinit.exe
(at least in the RemoteApp case) and the process are not allowed to escape it
therefore when the job assignment is attempted it failes with ERROR_PERMISSION_DENIED.
This is not a problem in Windows 8/Server 2012 because these allow nested jobs so we
should only respect this flag for versions older than that.

BUG=79091
TEST=Start Chrome as a published app with --allow-no-job and observe it spawning renderer processes properly.


Review URL: https://chromiumcodereview.appspot.com/10908171

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9c8b77de755451c12936655fd82f9f17b241702d