Seccomp-BPF: relax failure in probe process setup
When we set-up the probe process to test seccomp-bpf availability, setting
a pipe on stderr can sometimes fail. Presumably if this descriptor is backed
by a file on a file system that will return an error on close().

We don't consider not being able to set-up the pipe on stderr as a fatal error
anymore.

BUG=152530
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/11300014

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 224efd4b9542777ee030c4f972b52b9dbd6cc686