Windows: Allow subprocesses to inherit non-console stdout/stderr
Before this change, the renderer and other subprocesses never inherit
stdout and stderr when they are pipe handles.  Stdout/stderr will be
pipe handles when chrome.exe/browser_tests.exe is running under
Buildbot or under Cygwin's default terminal, mintty.

We fix this by specifying PROC_THREAD_ATTRIBUTE_HANDLE_LIST in the
arguments to CreateProcess().

The fix only applies on Windows >=Vista.

Although it's probably safe for stdout/stderr to be inherited when it
is a pipe handle or file handle, we put this behind the flag
"--enable-logging".  (This flag already makes stderr work when
chrome.exe/browser_tests.exe is running under a Windows console -- a
case which is not handled by the code path we're adding here because a
Windows console is not an inheritable kernel handle.)

Note that this relies on the fix committed in http://crrev.com/178656.

BUG=171836
TEST=manually add logging to renderer process and check that it appears
  when running chrome.exe or browser_tests.exe


Review URL: https://chromiumcodereview.appspot.com/12033045

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7d80efdebd6fdea4d9bd5eb445afe952f2b48598