sandbox_linux_unittests: ignore TSAN when counting threads.
TSAN uses a helper thread. For all intent and purposes it should be hidden
and it looks like we can fork() happily with this thread running.
So when counting threads, we ignore it.

We still disable BPF sandbox tests under TSAN though, the BPF sandbox is not
compatible with TSAN.

BUG=256905,243968
R=glider@chromium.org

Review URL: https://codereview.chromium.org/18603002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9bb49dff2b9b5a09bd51f49e942e68739f7032ab