Extract CreateInitProcessReaper() from the Zygote.
CreateInitProcessReaper() is useful for the CLONE_NEWUSER-based
sandbox as well as for the current Zygote. Extract it so that it
can be used independantly of content/.

CreateInitProcessReaper() is now exposed to content through the
setuid sandbox client.

BUG=312380

Review URL: https://codereview.chromium.org/90243002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 97692ca9d50f8cbf31c56fdf53be9daf5d9e3ac8