Making sandbox error message on moving to namespaces more accurate.
When trying to move to new namespaces if a feature is supported but doesn't
work we fail with an error message.

The old error message was always "Failed to move to new PID namespace" but the
cause could be a failure to move to a new "network namespace".

BUG=326039

Review URL: https://codereview.chromium.org/102723004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 888c941ef3ac3dea975119d30163d6b6e2b11b9d