Reland: Add seccomp sandbox for non-SFI NaCl
This is the reland of https://codereview.chromium.org/196793023/
In the old patch, ldflags! for ASan was specified in nacl_loader target,
which is a static_library. Now we set this in nacl_helper target. The
diff from the previous change is:
https://codereview.chromium.org/240783003/

All syscalls except whitelisted ones will cause SIGSYS.

We test the sandbox with BPF_TEST and BPF_TEST_DEATH, which
appropriately fork the process so the main process of the test
will never enable the sandbox.

TEST=Our app works with this sandbox on i686 and ARM
TEST=Build chrome and nacl_helper on i686, x86-64, and ARM
TEST=./out/Release/components_unittests --gtest_filter='NaClNonSfi*'
TEST=SFI NaCl apps still work
TEST=trybots
BUG=359285
R=jln@chromium.org, mseaborn@chromium.org
TBR=jochen

Review URL: https://codereview.chromium.org/239703011

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 13d80f03a33ccb9a1db015827bda9257eabe881a