NaCl non-sfi sandbox: use modern framework for tests
The non-sfi sandbox policy is the implementation of the SandboxBPFPolicy
interface but tests are testing it as an old-style function pointer
policy.
This CL makes use of BPF_TEST_C and BPF_DEATH_TEST_C and takes a classname
as a parameter rather than a function pointer.

BUG=359285
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/269333004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5a3e520c8db793d414e574e0a74c61ff27dacd7d