Replace NT prefix in sandbox rules match string to handle correct wildcard escaping
This patch adds a function to modify file system sandbox rules to replace the
\??\ NT prefix with the correct escaped form \/?/?\ for the wildcard matching
rules in the broker. This is done generally as it's a common mistake in the
sandbox code and so provides some defence in depth.

BUG=334882

Review URL: https://codereview.chromium.org/432543005

Cr-Original-Commit-Position: refs/heads/master@{#290131}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7b48bebc27bd4e18db785b020b07242977d0ca05