Linux sandbox: refactor BrokerProcess class
The BrokerProcess class is too complex. It both serves as management
of a new broker process, as well as the implementation of an
async-signal-safe IPC mechanism suitable to broker syscall.

We split the BrokerProcess class into 4 simpler classes. This will make it
easier to replace the broker process with a thread or a new Chrome process
type.

BUG=429028

Review URL: https://codereview.chromium.org/688843003

Cr-Original-Commit-Position: refs/heads/master@{#302370}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 70c42c20bd408c9f1b4d43f5154da43c16529cf8