CodeGen: refactor API [1/3]
In preparation for a rewrite of CodeGen, refactor the API slightly.
In particular, instead of making the API in terms of "Instruction*",
introduce a typedef "CodeGen::Node" so we can change the underlying
type later.

Also, remove the MakeInstruction() overload in favor of a single entry
point with a variable length parameter list.

BUG=414363

Review URL: https://codereview.chromium.org/703533002

Cr-Original-Commit-Position: refs/heads/master@{#304509}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d39e72c75aaa3c75b99907350d9eb69f82b7ce7b