Added a new process mitigation to harden process token IL policy.
This adds a new process mitigation policy to harden the current process
token's integrity level policy. What this actually means is the token's
IL policy in its SACL is modified to add no-read-up and no-execute-up
which is not the default. This prevents a lower privilege process from
opening the token object with rights such as duplicate and impersonation
which could be used to circumvent sandbox restrictions and elevate
privileges. While the policy is only enabled on the browser process by
making it a general mitigation policy it could be applied to all process
levels such as the GPU process to provide a similar effect.

BUG=440692

Review URL: https://codereview.chromium.org/810083002

Cr-Original-Commit-Position: refs/heads/master@{#313099}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 19669894e93f9279e860d9fff6a54f6cd042acd7