bpf_dsl: decouple PolicyCompiler from Syscall
Logically, the "escape hatch PC" is now like a compiler flag, that the
user can set appropriately.  In the case of SandboxBPF, it will set the
PC to Syscall::Call(-1), as before.

This isn't a very satisfying way to resolve this cyclic dependency, but
it's the simplest and least intrusive I could think of.

BUG=449357

Review URL: https://codereview.chromium.org/939943002

Cr-Original-Commit-Position: refs/heads/master@{#317237}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 242ad8711d9f0d58fb45b134468ad68b0835fad2