Fix some clang warnings with -Wmissing-braces in sandbox.
Clang warns if there are missing braces around a subobject
initializer. The most common idiom that triggers this is:
  STRUCT s = {0};
if the first field of STRUCT is itself a struct. This can
be more simply written as:
  STRUCT s = {};
which also prevents the warning from firing.

BUG=505297

Review URL: https://codereview.chromium.org/1216243007

Cr-Original-Commit-Position: refs/heads/master@{#337237}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9d02fb0d79f255dbaf8e8613ee829d8875ae561d