ThreadHelpers unit tests were racy
ThreadHelper checks /proc/self to count threads, to know a thread
exists in the process. But, /proc would be updated lately.
For thread.Stop(), there is a mechanisms to stop the thread and wait for
/proc being updated, but there isn't for thread.Start().

This will make the test flaky if we change the thread implementation.

BUG=468793

Review URL: https://codereview.chromium.org/1233823005

Cr-Original-Commit-Position: refs/heads/master@{#340857}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 15d8851020a8684b76f5d90a9dd1277ed110391b