Change type of kSUIDSandboxApiNumber to int
kSUIDSandboxApiNumber is declared as long, but treated as int in C++
code. C code treats it as a long, apparently in order to be able to use
strtol.

Declare it and treat it consistently as an int.

TEST=sandbox_linux_unittests
BUG=505479

Review URL: https://codereview.chromium.org/1348933003

Cr-Original-Commit-Position: refs/heads/master@{#350096}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 58ceb59e54fafb2cfda634d0966d1be056225515